Force unaligned load/store of widen vectors
Renderscript front-end compiles non-power-2 vector with additional
alignment in LLVM IR. It is incorrect as Renderscript only assume
element-alignement for vector types in load/store according to:
http://developer.android.com/reference/android/renderscript/Element.html

This workaround resets non-power-2 widen-vector's alignment specified
by frontend to 1 and forces load/store of widen vectors into unaligned one.
This results in unaligned instruction be used, e.g. movups instead of movaps.

In addition, a previous clang bug workaround of packed structure issue
is removed as this workaround covers it as well.

Change-Id: I82768d196968ece96710bfdcef0bca6b05355190
Author: Michael Liao <michael.liao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: Yong Chen <yong.a.chen@intel.com>